requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
tweepy>=4.0.0
textblob>=0.17.1
fastapi-limiter>=1.1.0
redis>=4.3.4